[turbofan] Fix Projection operator parameter type.
R=svenpanne@chromium.org

Review URL: https://codereview.chromium.org/549063002

git-svn-id: https://v8.googlecode.com/svn/branches/bleeding_edge@23764 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ src/compiler/common-operator.h
@@ -114,8 +114,9 @@ class CommonOperatorBuilder {
                                       1, "Parameter", index);
   }
   Operator* Int32Constant(int32_t value) {
-    return new (zone_) Operator1<int>(IrOpcode::kInt32Constant, Operator::kPure,
-                                      0, 1, "Int32Constant", value);
+    return new (zone_)
+        Operator1<int32_t>(IrOpcode::kInt32Constant, Operator::kPure, 0, 1,
+                           "Int32Constant", value);
   }
   Operator* Int64Constant(int64_t value) {
     return new (zone_)
@@ -177,9 +178,9 @@ class CommonOperatorBuilder {
   Operator* Call(CallDescriptor* descriptor) {
     return new (zone_) CallOperator(descriptor, "Call");
   }
-  Operator* Projection(int index) {
-    return new (zone_) Operator1<int>(IrOpcode::kProjection, Operator::kPure, 1,
-                                      1, "Projection", index);
+  Operator* Projection(size_t index) {
+    return new (zone_) Operator1<size_t>(IrOpcode::kProjection, Operator::kPure,
+                                         1, 1, "Projection", index);
   }
 
  private:
